+++ app/Http/Controllers/PermissionController.php
@@ -13,16 +13,35 @@ use Spatie\Permission\Models\Permission;
 class PermissionController extends Controller
 {
     public function index (){
+
+        if(!auth()->user()->can('View_Permission')){
+            abort(404);
+        }
         return view('permission.index');
     }
 
     public function ssd (Request $request){
+
+        if(!auth()->user()->can('View_Permission')){
+            abort(404);
+        }
+
         $permissions = Permission::query();
         return Datatables::of($permissions)
     
         ->addColumn('action',function($each){
-            $edit_icon = '<a href=" '. route('permission.edit',$each->id) .'  " class="text-warning me-2"><i class="fas fa-edit"></i></a>';
-            $delete_icon = '<a href="#" class="text-danger delete-btn ms-3" data-id=" '.$each->id.' " ><i class="fas fa-trash-alt"></i></a>';
+
+            $edit_icon = '';
+            $delete_icon = '';
+
+
+            if(auth()->user()->can('Edit_Permission')){
+                $edit_icon = '<a href=" '. route('permission.edit',$each->id) .'  " class="text-warning me-2"><i class="fas fa-edit"></i></a>';
+            }
+
+            if(auth()->user()->can('Delete_Permission')){
+                $delete_icon = '<a href="#" class="text-danger delete-btn ms-3" data-id=" '.$each->id.' " ><i class="fas fa-trash-alt"></i></a>';
+            }
          
 
             
@@ -38,10 +57,17 @@ class PermissionController extends Controller
     // create Department page
     public function create () {
 
+        if(!auth()->user()->can('Create_Permission')){
+            abort(404);
+        }
         return view ('permission.create');
     }
 
     public function store (StorePermission $request){
+
+        if(!auth()->user()->can('Create_Permission')){
+            abort(404);
+        }
 
         $permission = new Permission();
         $permission->name = $request->name;
@@ -52,12 +78,20 @@ class PermissionController extends Controller
 
     // edit page 
     public function edit ($id) {
+
+        if(!auth()->user()->can('Edit_Permission')){
+            abort(404);
+        }
         $permissions = Permission::findOrFail($id);
         return view('permission.edit',compact('permissions'));
     }
 
     // update 
     public function update ($id, UpdatePermission $request){
+
+        if(!auth()->user()->can('Edit_Permission')){
+            abort(404);
+        }
             $permission = Permission::findOrFail($id);
             $permission->name = $request->name;
             $permission->update();
@@ -67,6 +101,10 @@ class PermissionController extends Controller
 
     // department delete
     public function destroy ($id){
+
+        if(!auth()->user()->can('Delete_Permission')){
+            abort(404);
+        }
         $permission = Permission::findOrFail($id);
         $permission->delete();
