+++ app/Http/Controllers/RoleController.php
@@ -16,10 +16,18 @@ use Spatie\Permission\Models\Permission;
 class RoleController extends Controller
 {
     public function index (){
+
+        if(!auth()->user()->can('View_Role')){
+            abort(404);
+        }
         return view('role.index');
     }
 
     public function ssd (Request $request){
+
+        if(!auth()->user()->can('View_Role')){
+            abort(404);
+        }
 
         $departments = Role::query();
         return Datatables::of($departments)
@@ -32,8 +40,18 @@ class RoleController extends Controller
             return $output;
         })
         ->addColumn('action',function($each){
-            $edit_icon = '<a href=" '. route('role.edit',$each->id) .'  " class="text-warning me-2"><i class="fas fa-edit"></i></a>';
-            $delete_icon = '<a href="#" class="text-danger delete-btn ms-3" data-id=" '.$each->id.' " ><i class="fas fa-trash-alt"></i></a>';
+
+            $edit_icon = '';
+            $delete_icon = '';
+
+
+            if(auth()->user()->can('Edit_Role')){
+                $edit_icon = '<a href=" '. route('role.edit',$each->id) .'  " class="text-warning me-2"><i class="fas fa-edit"></i></a>';
+            }
+
+            if(auth()->user()->can('Delete_Role')){
+                $delete_icon = '<a href="#" class="text-danger delete-btn ms-3" data-id=" '.$each->id.' " ><i class="fas fa-trash-alt"></i></a>';
+            }
             
             return '<span class="action-icon">'. $edit_icon . $delete_icon .'</span>';
         })
@@ -47,6 +65,9 @@ class RoleController extends Controller
     // create Department page
     public function create () {
 
+        if(!auth()->user()->can('Create_Role')){
+            abort(404);
+        }
         $permissions = Permission::all();
         return view ('role.create',compact('permissions'));
 
@@ -54,6 +75,10 @@ class RoleController extends Controller
     }
 
     public function store (StoreRole $request){
+
+        if(!auth()->user()->can('Create_Role')){
+            abort(404);
+        }
 
         $role = new Role();
         $role->name = $request->name;
@@ -67,6 +92,9 @@ class RoleController extends Controller
     // edit page 
     public function edit ($id) {
 
+        if(!auth()->user()->can('Edit_Role')){
+            abort(404);
+        }
         $role = Role::findOrFail($id);
         $permissions = Permission::all();
         $old_permission = $role->permissions->pluck('id')->toArray();
@@ -75,6 +103,10 @@ class RoleController extends Controller
 
     // update 
     public function update ($id, UpdateRole $request){
+
+        if(!auth()->user()->can('Edit_Role')){
+            abort(404);
+        }
             $role = Role::findOrFail($id);
             $role->name = $request->name;
             $role->update();
@@ -88,6 +120,10 @@ class RoleController extends Controller
 
     // department delete
     public function destroy ($id){
+
+        if(!auth()->user()->can('Delete_Role')){
+            abort(404);
+        }
         $role = Role::findOrFail($id);
         $role->delete();
